==> 1. Transform.Rmd <==
---
title: "Transforming raw data"
author: "Evgenii Berdinskikh, Marina Kozyreva, Anton Tikhonenko"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}

knitr::opts_chunk$set(
  echo = TRUE,
  fig.path = "data/pics/transform-"
  )

library(conflicted)
library(readxl)
library(skimr)
library(tidyverse)
library(zoo)

conflicts_prefer(dplyr::filter)
conflicts_prefer(dplyr::select)
conflicts_prefer(dplyr::lag)

Sys.setlocale("LC_TIME", "C") %>% invisible()

```

## Data Preparation

```{r read, message=FALSE, warning=FALSE}

# Ambulance Data in Irkutsk (2006 - 2022)
brain_data_raw <- read_excel("data/ems/database_2024-11-09.xlsx") ## Acute cerebrovascular accident
heart_data_raw <- read_excel("data/ems/database_2025-01-18.xlsx") ## Heart diseases

# Geomagnetic activity indices (Ap and Kp) from GFZ, Potsdam (2006 - 2022)
GFZ_Potsdam_Ap <- read_csv("data/raw/GFZ_Potsdam_Ap.csv")  
GFZ_Potsdam_Kp <- read_csv("data/raw/GFZ_Potsdam_Kp.csv") %>% select(Date, Kp_Sum)   

# Geomagnetic activity index (Dst) from WDC, Kyoto (2006 - 2022)
WDC_Kyoto_Dst <- 
  read_fwf("data/raw/WDC_Kyoto.dat", 
           fwf_widths(c(3, 2, 2, 1, 2, 2, 1, 1, 2, 4, rep(4, 24), 4)), 
           show_col_types = FALSE) %>% 
  rowwise() %>%
  transmute(
    Date = as.Date(paste0(X9, X2, "-", X3, "-", X5)), Dst_mean = X35, ## Daily average Dst value
    Dst_max = max(c_across(X11:X34), na.rm = TRUE), Dst_min = min(c_across(X11:X34), na.rm = TRUE), 
    Dst_var = Dst_max - Dst_min,  ## Daily Dst value amplitude
    Dst_level = factor(case_when( ## Geomagnetic activity level`
      Dst_min < -150 ~ "Severe", 
      Dst_min < -100 ~ "Major", 
      Dst_min < -50 ~ "Moderate",  
      Dst_min < -20 ~ "Weak", 
      TRUE ~ "Quiet"), levels = c("Quiet", "Weak", "Moderate", "Major", "Severe"), ordered = TRUE) 
    )  %>% 
  ungroup() %>% 
  select(-Dst_min, -Dst_max)

## `Dimitrov S, Mustafa FR, Stoilova I, Babayev ES, Kazimov EA (2009) Possible influence of solar extreme events and related geomagnetic disturbances on human cardio-vascular state: results of collaborative Bulgarian-Azerbaijani studies. Adv Space Res 43:641–648. https://doi.org/10.1016/j.asr.2008.09.006

# Data on sudden events: SSC (Sudden Storm Commencement) and SI (Sudden Impulse) from Observatori de l'Ebre (2006 - 2022)
Obs_Ebre_SSC <- tibble()

for (i in (2006:2022)) {
  link <- paste0("data/raw/SC_data/ssc_", i, "_d.txt")
  
  Obs_Ebre_SSC %>% rbind(
    
    read_tsv(link, col_names = "raw") %>%
      filter(str_starts(raw, paste0(i))) %>%
      transmute(
        Date = str_replace_all(str_trunc( raw, width = 10, ellipsis = ""), " ", "-") %>% as.Date(),
        Sudden_Start = str_remove_all(raw, "-") %>% 
          str_trunc( width = 2, side = "left", ellipsis = "" ))
    
    ) -> Obs_Ebre_SSC
  }

Obs_Ebre_SSC <- tibble(
  Date = seq.Date(from = as.Date("2006-01-01"), to = as.Date("2022-12-31"), by = "day")) %>% 
  left_join(
    Obs_Ebre_SSC %>% group_by(Date) %>% summarize(Sudden_Start = n(), .groups = "drop"), 
    by = join_by(Date)) %>%
  mutate(Sudden_Start = ifelse(is.na(Sudden_Start), "No", "Yes") %>% as.factor()) 
  
# Weather data in Irkutsk from Rosgidromet (1882 - 2023)
Roshydromet_Irkutsk <- read_fwf(
  "data/raw/Roshydromet_Irkutsk.dat",
  fwf_widths(c( 5, 1, 4, 1, 2, 1, 2, 1, 1, 1, 5, 1, 1, 1, 5, 1, 1, 1, 5, 1, 1, 1, 5, 1, 1, 1))) %>%
  select(where(is.numeric)) %>%
  transmute(
    Date = as.Date(paste0(X3, "-", X5, "-", X7)),
    Temp_mean = X15,  ## Daily temperature (average, minimum, maximum)
    Temp_min = X11,
    Temp_max = X19,
    H2O = X23 ## Daily precipitation
    ) 

# International sunspot number from SIDC, Brussels (1818 - 2024)
SIDC_SILSO <- read_delim("data/raw/SIDC_SILSO.csv", delim = ";", col_names = FALSE)  %>% 
  transmute(Date = as.Date(paste0(X1, "-", X2, "-", X3)), 
            Sunspots = as.numeric(X5))

# Data on local K index from the Irkutsk magnetic observatory (IRT)
K_IRT <- read_excel("data/raw/K_IRT/IRT_K_index.xlsx", skip = 5) 

K_IRT %>% 
  transmute(Date = as.Date(`DA-MON-YR`, format = "%d-%b-%y"), 
            K_00 = `0`, K_03 = `0.125`, K_06 = `0.25`, K_09 = `0.375`, 
            K_12 = `0.5`, K_15 = `0.625`, K_18 = `0.75`, K_21 = `0.875`, 
            K_Sum = SK) %>%  
  mutate(across(starts_with("K"), ~ifelse(. == -1, NA, .)) ) %>%  
  rbind(
    tibble(Date = seq.Date(from = as.Date("2021-01-01"), to = as.Date("2022-12-31"), by = "day")) %>% 
    left_join( rbind( 
      read_table("data/raw/K_IRT/IRT2021.txt", show_col_types = FALSE, col_names = FALSE, skip = 1) %>%  
        distinct() %>% filter(!is.na(X3)), 
      read_table("data/raw/K_IRT/IRT2022.txt", show_col_types = FALSE, col_names = FALSE, skip = 1) %>%  
        distinct() %>% filter(!is.na(X3))
      ) %>% 
        transmute( Date = X1, Point = paste0("K_", substr(X2, 1, 2)), 
                   K_IRT = as.numeric(gsub(";", "", X4)) ) %>%  
        pivot_wider( names_from = Point, values_from = K_IRT ) %>% 
        rowwise() %>% 
        mutate( across(starts_with("K"), ~ifelse(. == -1, NA, .)), 
          K_Sum = sum(c_across(starts_with("K")), na.rm = FALSE) ) %>% 
        ungroup(), 
      by = join_by(Date)) 
    ) %>% 
  select(Date, K_Sum) -> K_IRT
    
# Production calendar in Russia (2006 - 2022)
Calendar <- read_csv("data/raw/Calendar.csv") %>% 
  mutate(across(c(Season, Month, DayOff), as.factor))

```

## Data Merging

```{r join}

# Ambulance Data related to acute cerebrovascular accidents
brain_data <- brain_data_raw %>% 
  transmute(
    Date = dmy(Дата), EMS = `Вызов СМП с диагнозом ОНМК у всех пациентов старше 25 лет`, 
    ## Missing data are replaced with the mean of neighboring values
    F10.7 = `Cкорректированный (F10.7)` %>% na.approx(na.rm = FALSE), 
    Wind = ifelse(is.na(`Скорость ветра`), na.approx(`Скорость ветра`, na.rm = FALSE), `Скорость ветра`), 
    Pressure = `Атмосферное давление в среднем за сутки, С (Т)` %>%  
      ifelse(. < 600, NA, .) %>% na.approx(na.rm = FALSE) ) %>% 
  filter(!is.na(Date)) %>%
  left_join(Calendar, by = join_by(Date)) %>%  ## Calendar data
  left_join(GFZ_Potsdam_Kp, by = join_by(Date)) %>%  ## Planetary K-index
  left_join(GFZ_Potsdam_Ap, by = join_by(Date)) %>%  ## Ap Index
  left_join(K_IRT, by = join_by(Date)) %>%  ## IRT K-index
  left_join(Obs_Ebre_SSC, by = join_by(Date)) %>%  ## Sudden Impulse/Storm Commencement
  left_join(Roshydromet_Irkutsk, by = join_by(Date)) %>%  ## Temperature & Precipitation
  left_join(SIDC_SILSO, by = join_by(Date)) %>%   ## Sunspots
  left_join(WDC_Kyoto_Dst, by = join_by(Date)) %>%  ## Disturbance Storm-Time (Dst) 
  mutate(COVID = as.factor(ifelse(Date >= "2020-03-27", "Yes", "No"))) %>% 
  select(Date, Year, Season, Month, DayOff, COVID, Temp_mean, Temp_min, Temp_max, H2O, Wind, Pressure,
         Sunspots, F10.7, Sudden_Start, Kp_Sum, Ap_mean, K_Sum,  Dst_mean, Dst_var, Dst_level, EMS)

# Ambulance Data related to cardiovascular diseases
heart_data <- brain_data %>% 
  select(!EMS) %>% 
  left_join(
    heart_data_raw %>% 
      transmute( 
        Date = seq.Date(from = as.Date("2006-11-29"), by = "day", length.out = n()), 
        EMS_UA = `Вызова СМП с диагнозом НС у всех пациентов старше 25 лет`,  ## Unstable angina
        EMS_AMI = `Вызова СМП с диагнозом ОИМ у всех пациентов старше 25 лет`), ## Acute Myocardial Infarction
    by = join_by(Date)) 

```

## Pivot Tables

```{r long}

# Ambulance Data related to acute cerebrovascular accidents
brain_data_long <- brain_data_raw %>% 
  transmute(
    Date = dmy(Дата),
    ## Ambulance data by gender and age
    EMS_Male_1 = `Количество вызовов СМП с диагнозом ОНМК у мужчин 25-44 лет`,
    EMS_Male_2 = `Количество вызовов СМП с диагнозом ОНМК у мужчин 45-54 лет`,
    EMS_Male_3 = `Количество вызовов СМП с диагнозом ОНМК у мужчин старше 55 лет`,
    EMS_Female_1 = `Количество вызовов СМП с диагнозом ОНМК у женщин 25-44 лет`,
    EMS_Female_2 = `Количество вызовов СМП с диагнозом ОНМК у женщин 45-54 лет`,
    EMS_Female_3 = `Количество вызовов СМП с диагнозом ОНМК у женщин старше 55 лет`,
    ## Missing data are replaced with the mean of neighboring values
    F10.7 = `Cкорректированный (F10.7)` %>% na.approx(na.rm = FALSE), 
    Wind = ifelse(is.na(`Скорость ветра`), na.approx(`Скорость ветра`, na.rm = FALSE), `Скорость ветра`), 
    Pressure = `Атмосферное давление в среднем за сутки, С (Т)` %>%  
      ifelse(. < 600, NA, .) %>% na.approx(na.rm = FALSE) ) %>% 
  filter(!is.na(Date)) %>%
  left_join(Calendar, by = join_by(Date)) %>%  ## Calendar data
  left_join(GFZ_Potsdam_Kp, by = join_by(Date)) %>%  ## Planetary K-index
  left_join(GFZ_Potsdam_Ap, by = join_by(Date)) %>%  ## Ap Index
  left_join(K_IRT, by = join_by(Date)) %>%  ## IRT K-index
  left_join(Obs_Ebre_SSC, by = join_by(Date)) %>%  ## Sudden Impulse/Storm Commencement
  left_join(Roshydromet_Irkutsk, by = join_by(Date)) %>%  ## Temperature & Precipitation
  left_join(SIDC_SILSO, by = join_by(Date)) %>%   ## Sunspots
  left_join(WDC_Kyoto_Dst, by = join_by(Date)) %>%  ## Disturbance Storm-Time (Dst) 
  mutate(COVID = as.factor(ifelse(Date >= "2020-03-27", "Yes", "No"))) %>% 
  select(Date, Year, Season, Month, DayOff, COVID, Temp_mean, Temp_min, Temp_max, H2O, Wind, Pressure,
         Sunspots, F10.7, Sudden_Start, Kp_Sum, Ap_mean, K_Sum,  Dst_mean, Dst_var, Dst_level, 
         EMS_Male_1, EMS_Male_2, EMS_Male_3, EMS_Female_1, EMS_Female_2, EMS_Female_3) %>% 
  ## Long data format, 1 day - 6 rows
  pivot_longer(cols = starts_with("EMS_"), 
               names_to = c("Sex", "Age"), 
               names_pattern = "EMS_(Male|Female)_(1|2|3)", 
               values_to = "EMS") %>% 
  mutate(Sex = as.factor(Sex), 
         Age = as.factor(case_when(
         Age == 1 ~ "25-44", Age == 2 ~ "45-54", Age == 3 ~ "55+"))) 


# Ambulance Data  related to unstable angina
UA_data_long <- brain_data %>% 
  select(!EMS) %>% 
  left_join(
    heart_data_raw %>% 
      transmute( 
        Date = seq.Date(from = as.Date("2006-11-29"), by = "day", length.out = n()), 
        ## Ambulance data by gender and age
        EMS_Male_1 = `Количество вызовов СМП с диагнозом НС у мужчин 25-44 лет`,
        EMS_Male_2 = `Количество вызовов СМП с диагнозом НС у мужчин 45-54 лет`,
        EMS_Male_3 = `Количество вызовов СМП с диагнозом НС у мужчин старше 55 лет`,
        EMS_Female_1 = `Количество вызовов СМП с диагнозом НС у женщин 25-44 лет`,
        EMS_Female_2 = `Количество вызовов СМП с диагнозом НС у женщин 45-54 лет`,
        EMS_Female_3 = `Количество вызовов СМП с диагнозом НС у женщин старше 55 лет`),
    by = join_by(Date)) %>% 
  ## Long data format, 1 day - 6 rows
  pivot_longer(cols = starts_with("EMS_"), 
               names_to = c("Sex", "Age"), 
               names_pattern = "EMS_(Male|Female)_(1|2|3)", 
               values_to = "EMS") %>% 
  mutate(Sex = as.factor(Sex), 
         Age = as.factor(case_when(
         Age == 1 ~ "25-44", Age == 2 ~ "45-54", Age == 3 ~ "55+"))) 


# Ambulance Data related to acute myocardial infarction
AMI_data_long <- brain_data %>% 
  select(!EMS) %>% 
  left_join(
    heart_data_raw %>% 
      transmute( 
        Date = seq.Date(from = as.Date("2006-11-29"), by = "day", length.out = n()), 
        ## Ambulance data by gender and age
        EMS_Male_1 = `Количество вызовов СМП с диагнозом ОИМ у мужчин 25-44 лет`,
        EMS_Male_2 = `Количество вызовов СМП с диагнозом ОИМ у мужчин 45-54 лет`,
        EMS_Male_3 = `Количество вызовов СМП с диагнозом ОИМ у мужчин старше 55 лет`,
        EMS_Female_1 = `Количество вызовов СМП с диагнозом ОИМ у женщин 25-44 лет`,
        EMS_Female_2 = `Количество вызовов СМП с диагнозом ОИМ у женщин 45-54 лет`,
        EMS_Female_3 = `Количество вызовов СМП с диагнозом ОИМ у женщин старше 55 лет` ),
    by = join_by(Date)) %>% 
  ## Long data format, 1 day - 6 rows
  pivot_longer(cols = starts_with("EMS_"), 
               names_to = c("Sex", "Age"), 
               names_pattern = "EMS_(Male|Female)_(1|2|3)", 
               values_to = "EMS") %>% 
  mutate(Sex = as.factor(Sex), 
         Age = as.factor(case_when(
         Age == 1 ~ "25-44", Age == 2 ~ "45-54", Age == 3 ~ "55+"))) 

```

## Overview of Tables

```{r summary}

skim(brain_data)
skim(brain_data_long)

```

## Files Reading and Writing

```{r write, message=FALSE}

#Local data storage 
write_csv(brain_data, "data/ems/brain_data.csv") 
write_csv(heart_data, "data/ems/heart_data.csv") 
write_csv(brain_data_long, "data/ems/brain_data_long.csv") 
write_csv(UA_data_long, "data/ems/UA_data_long.csv") 
write_csv(AMI_data_long, "data/ems/AMI_data_long.csv") 

#Import data
## Ambulance Data related to acute cerebrovascular accidents
brain_data <- read_csv("data/ems/brain_data.csv") %>%
   mutate(across(c(Year, Season, Month, DayOff, COVID, Sudden_Start, Dst_level), as.factor))
## Ambulance Long Data related to acute cerebrovascular accidents
brain_data_long <- read_csv("data/ems/brain_data_long.csv") %>%
    mutate(across(c(Year, Season, Month, DayOff, COVID, Sudden_Start, Dst_level, Sex, Age), as.factor))

## Ambulance Data related to cardiovascular diseases
heart_data <- read_csv("data/ems/heart_data.csv") %>%
   mutate(across(c(Year, Season, Month, DayOff, COVID, Sudden_Start, Dst_level), as.factor))
## Ambulance Long Data  related to unstable angina
UA_data_long <- read_csv("data/ems/UA_data_long.csv") %>%
   mutate(across(c(Year, Season, Month, DayOff, COVID, Sudden_Start, Dst_level, Sex, Age), as.factor))
## Ambulance Long Data related to acute myocardial infarction
AMI_data_long <- read_csv("data/ems/AMI_data_long.csv") %>%
   mutate(across(c(Year, Season, Month, DayOff, COVID, Sudden_Start, Dst_level, Sex, Age), as.factor))

```
