==> 3. Models.Rmd <==
---
title: "Regression analysis"
author: "Evgenii Berdinskikh, Marina Kozyreva, Anton Tikhonenko"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}

knitr::opts_chunk$set(
  echo = TRUE,
  fig.path = "data/pics/models-"
  )

library(AER)
library(broom)
library(car)
library(conflicted)
library(flextable)
library(ggfortify)
library(ggpubr)
library(gt)
library(MASS)
library(tidyverse)

conflicts_prefer(dplyr::filter)
conflicts_prefer(dplyr::select)
conflicts_prefer(dplyr::lag)

Sys.setlocale("LC_TIME", "en_US.UTF-8") %>% invisible()

theme_custom <- theme_bw()+ theme(
    plot.title = element_text(size = 30, hjust = 0.5),
    plot.subtitle = element_text(size = 25, hjust = 0.5),
    strip.text = element_text(size = 20),
    axis.text = element_text(size = 20),
    axis.title = element_text(size = 25),
    legend.title = element_text(size = 25),
    legend.text = element_text(size = 20)
  )

```

## Data import

```{r read, message=FALSE}

## Ambulance Data related to acute cerebrovascular accidents
brain_data <- read_csv("data/ems/brain_data.csv") %>%
   mutate(across(c(Year, Season, Month, DayOff, COVID, Sudden_Start, Dst_level), as.factor))
## Ambulance Long Data related to acute cerebrovascular accidents
brain_data_long <- read_csv("data/ems/brain_data_long.csv") %>%
    mutate(across(c(Year, Season, Month, DayOff, COVID, Sudden_Start, Dst_level, Sex, Age), as.factor))

```

## Basic model

The distribution of ambulance calls follows a Poisson distribution, as observed in the graph:

```{r distribution, fig.height=8, fig.width=16}

brain_data %>%
  ggplot(aes(x = EMS, after_stat(density) )) +
  geom_histogram(binwidth = 4, fill = "deepskyblue", colour = "deepskyblue4", alpha = 0.5) +
  geom_density(color = "tomato2", linewidth = 2) +
  theme_custom +
  labs(x = "Number of Emergency Calls", 
       y = "Density", 
       title = "Distribution of Emergency Calls") +
  theme_custom

```

The baseline Poisson regression model included the date, average daily temperature, number of sunspots, and the total Kp index for the day. All coefficients are significant, p \< 0.05.

```{r model}

poisson_m0 <- glm(EMS ~ Date + Temp_mean + Sunspots + Kp_Sum,  
    data = brain_data, family = poisson)  
  
poisson_m0 %>% summary()

```

## Basic model diagnostic

The diagnostic plots and component-plus-residual plots do not reveal any major violations in the model, confirming its adequacy.

```{r plus, fig.height=10, fig.width=20}
autoplot(poisson_m0)+theme_custom
crPlots(poisson_m0)
```

However, overdispersion **(1.54)** and autocorrelation of residuals are present, which may indicate the need for model refinement.

```{r minus}

# Overdispersion 
dispersiontest(poisson_m0) 

# Autocorrelation of residuals
residuals(poisson_m0, type = "response") %>% acf() 

```

## Alternative models

As an alternative, we considered negative binomial regression and the use of robust standard errors for the Poisson regression.

```{r alter, fig.width=14}

# Poisson regression
tidy(poisson_m0, conf.level = 0.95, conf.int = TRUE) %>%
  mutate(
    model = "Poisson",  
    EMS_effect = ifelse(
    term != "(Intercept)",
    paste0(signif((exp(estimate) - 1) * 100, 3), "%"),
    signif((exp(estimate)), 3))) -> Poisson_data

# Heteroscedasticity-consistent estimation  (vcovHC)
bind_cols(
  tidy(coeftest(poisson_m0, vcov. = vcovHC(poisson_m0, type = "HC3"))),
  coefci(poisson_m0, vcov. = vcovHC(poisson_m0, type = "HC3")) %>% 
    as_tibble() %>% rename(conf.low = "2.5 %", conf.high = "97.5 %")) %>%
  mutate(
    model = "Poisson_vcovHC", 
    EMS_effect = ifelse(
    term != "(Intercept)",
    paste0(signif((exp(estimate) - 1) * 100, 3), "%"),
    signif((exp(estimate)), 3))) -> Poisson_vcovHC_data

# Heteroscedasticity and Autocorrelation Consistent estimation (vcovHAC)
bind_cols(
  tidy(coeftest(poisson_m0, vcov. = vcovHAC(poisson_m0))),
  coefci(poisson_m0, vcov. = vcovHAC(poisson_m0)) %>% 
    as_tibble() %>% rename(conf.low = "2.5 %", conf.high = "97.5 %")) %>%
  mutate(
    model = "Poisson_vcovHAC", 
    EMS_effect = ifelse(
    term != "(Intercept)",
    paste0(signif((exp(estimate) - 1) * 100, 3), "%"),
    signif((exp(estimate)), 3))) -> Poisson_vcovHAC_data

# Negative binomial regression
NBR_model <- glm.nb(EMS ~ Date + Temp_mean + Sunspots + Kp_Sum,
                    data = brain_data, link = "log")

tidy(NBR_model, conf.level = 0.95, conf.int = TRUE) %>% 
  mutate(
    model = "Neg_binomial", 
    EMS_effect = ifelse(
    term != "(Intercept)",
    paste0(signif((exp(estimate) - 1) * 100, 3), "%"),
    signif((exp(estimate)), 3))) -> Neg_binomial_data

```

The use of robust standard errors (vcovHC/vcovHAC) and negative binomial regression significantly increases the confidence interval for the estimates. This further highlights the need to improve the Poisson model to address overdispersion and autocorrelation.

```{r compare, fig.height=5, fig.width=15}

bind_rows(Poisson_data, Poisson_vcovHC_data, Poisson_vcovHAC_data, Neg_binomial_data) %>% 
  filter(term != "(Intercept)") %>%
  mutate(term = fct_recode(term, `Sunspot Number` = "Sunspots", 
                           `Average Temperature` = "Temp_mean", 
                           `Average Kp index` = "Kp_Sum"), 
         model = fct_recode(model, `Poisson (vcovHC)` = "Poisson_vcovHC", 
                           `Poisson (vcovHAC)`= "Poisson_vcovHAC", 
                           `Negative Binomial Regression` = "Neg_binomial")) %>% 
  ggplot(aes(x = term,y = estimate, ymin = conf.low, ymax = conf.high, colour = model))+
  geom_pointrange(position = position_dodge(width = 0.6)) +
  geom_hline(yintercept = 0, linetype = "dashed", color = "red") + 
  scale_color_brewer(palette = "Dark2", direction = -1)+
  coord_flip() + 
  labs( x = "", y = "Coefficient Estimates", title = "Model Comparison", color = "Model") +
  theme_custom

```

## Week lag

According to the ACF, spikes occur with a 7-day periodicity. We plan to address this by adding the **day of the week**! The ANOVA shows that including the new term significantly improves the model's explanatory power. The lags have been aligned, but their overall reduction still requires further adjustment.

```{r lag7, fig.height=8, fig.width=16}

brain_data_transform <- brain_data %>% 
  mutate(Day_Week = factor(weekdays(Date), 
                           levels = c("Monday", "Tuesday", "Wednesday", "Thursday",
                                      "Friday", "Saturday", "Sunday"))) 

poisson_m1 <- glm(EMS ~ Date + Day_Week + Temp_mean + Sunspots + Kp_Sum,
                               data = brain_data_transform, family = poisson)

anova(poisson_m0, poisson_m1, test = "LR")

ggarrange (
residuals(poisson_m0, type = "response") %>% acf() %>% autoplot() + 
  ggtitle("Basic model") + theme_custom,  
residuals(poisson_m1, type = "response") %>% acf() %>% autoplot() + 
  ggtitle("Basic model + Day of the week") + theme_custom
)

```

## Age & Gender

```{r long}

brain_data_long_transform <- brain_data_long %>% 
  mutate(Day_Week = factor(weekdays(Date), 
                           levels = c("Monday", "Tuesday", "Wednesday", "Thursday",
                                      "Friday", "Saturday", "Sunday"))) 

poisson_m2 <- glm(EMS ~ Date + Day_Week + Temp_mean + Sunspots + Kp_Sum + Age + Sex,
                  data = brain_data_long_transform, family = poisson)

```

The inclusion of new variables **(gender and age)** in the model, using the long format data **(1 day = 6 rows)**, reduced overdispersion from 1.54 to 1.20. The ACF drops sharply, but spikes persist at every 6th row!

```{r poisson_m2, fig.height=8, fig.width=16}
# Overdispersion 
dispersiontest(poisson_m0) 
 dispersiontest(poisson_m2) 


# Autocorrelation of residuals
ggarrange (
residuals(poisson_m1, type = "response") %>% acf() %>% autoplot() + 
  ggtitle("Without Age & Gender") + theme_custom,  
residuals(poisson_m2, type = "response") %>% acf() %>% autoplot() + 
  ggtitle("With Age & Gender") + theme_custom
)


```

Replacing **(Age + Sex)** with their interaction **(Age \* Sex)** shows an improvement in the model, as indicated by the ANOVA. Additionally, the lags are now more evenly distributed.

```{r Age_Sex, fig.height=10, fig.width=25}

poisson_m3 <- glm(EMS ~ Date + Day_Week + Age*Sex + Temp_mean + Sunspots + Kp_Sum,
                               data = brain_data_long_transform, family = poisson)

anova(poisson_m2, poisson_m3)


# Autocorrelation of residuals
ggarrange (
residuals(poisson_m1, type = "response") %>% acf() %>% autoplot() +
  ggtitle("Without Age & Sex") + theme_custom,
residuals(poisson_m2, type = "response") %>% acf() %>% autoplot() +
  ggtitle("Age+Sex") + theme_custom,
residuals(poisson_m3, type = "response") %>% acf() %>% autoplot() + 
  ggtitle("Age*Sex") + theme_custom, nrow = 1
)

```

## Calendar effect

Adding variables for **DayOff** and **COVID** shows meaningful improvements in the model, as confirmed by the ANOVA.

```{r Calendar}

# workdays vs weekend
poisson_m4 <- glm(EMS ~ Date + Day_Week + DayOff + Age*Sex + 
                        Temp_mean + Sunspots + Kp_Sum,
                               data = brain_data_long_transform, family = poisson)

anova(poisson_m3, poisson_m4, test = "LR")

# COVID-19
poisson_m5 <- glm(EMS ~ Date + Day_Week + DayOff + COVID + Age*Sex + 
                        Temp_mean + Sunspots + Kp_Sum,
                               data = brain_data_long_transform, family = poisson)

anova(poisson_m4, poisson_m5, test = "LR")

```

## Cyclic component

The additional variable accounts for the cyclic nature of the data and synchronizes it with the temperature.

```{r sin, fig.height=8, fig.width=16}

brain_data_long_transform <- brain_data_long_transform %>% 
  mutate(Temp_scaled = scale(Temp_mean), 
         Sin_year = sin(2 * pi * (yday(Date) - 103) / 365.25))

# RMSE function
RMSE_function <- function(shift) {
  brain_data_long_transform %>%
    mutate(
      Sin_year = sin(2 * pi * (yday(Date) - shift) / 365.25),
      Sin_year_scaled = scale(Sin_year),
      Squared_Error = (Temp_scaled - Sin_year_scaled)^2 ) %>% 
    summarise(RMSE = sqrt(mean(Squared_Error, na.rm = TRUE))) %>% 
    pull(RMSE)
}

# Phase-shift search
shifts <- 90:120
RMSE_values <- sapply(shifts, RMSE_function)

tibble(Shift = shifts, RMSE = RMSE_values) %>%
  ggplot(aes(x = Shift, y = RMSE)) +
  geom_line() +
  geom_point() +
  labs(x = "Day Shift", y = "RMSE") +
  theme_custom +
  annotate("text", x = 103, y = 0.41, size = 10, 
           label = paste0("Optimal shift parameter:", shifts[which.min(RMSE_values)]))+ 
  annotate("text", x = 103, y = 0.40, size = 7, 
           label = paste0("RMSE:", round(min(RMSE_values), 3)))
  
  brain_data_long_transform %>%
    ggplot(aes(x = Date)) +
  geom_line(aes(y = Temp_scaled, color = "Temperature"), linewidth = 1) +
  geom_line(aes(y = Sin_year, color = "Cyclic component"), linewidth = 2)+
  scale_color_manual(values = c("Temperature" = "deepskyblue3", "Cyclic component" = "tomato3")) +
  scale_x_date(date_breaks = "2 year", date_labels = "%Y")+
  labs(x = "", y = "Annual Cycle", colour = "")+
  theme_custom
    
```

Adding the Cyclic Component, according to the ANOVA analysis, improves the model.

```{r cycle}

poisson_m6 <- glm(EMS ~ Date + Day_Week + DayOff + COVID + Sin_year + 
                        Temp_mean + Sunspots + Kp_Sum + Age*Sex,
                               data = brain_data_long_transform, family = poisson)

anova(poisson_m5, poisson_m6, test = "LR")

```

## Geomagnetic parameters

Replacing the daily Kp index with the **local K index** for Irkutsk reduces the AIC, confirming the choice of the local index. ANOVA also showed that adding the **Dst index** as the daily amplitude of values improves the model, while adding the presence of **Sudden Impulses** had no significant impact.

```{r Geomagnetic, warning=FALSE}

# K index (IRT)
poisson_m7 <- glm(EMS ~ Date + Day_Week + DayOff + COVID + Sin_year + 
                        Temp_mean + Sunspots + K_Sum + Age*Sex,
                               data = brain_data_long_transform, family = poisson)
AIC(poisson_m6, poisson_m7)

# Dst index (Dst_var)
poisson_m8 <- glm(EMS ~ Date + Day_Week + DayOff + COVID + Sin_year + 
                        Temp_mean + Sunspots + K_Sum + Dst_mean + Age*Sex,
                               data = brain_data_long_transform, family = poisson)
poisson_m9 <- glm(EMS ~ Date + Day_Week + DayOff + COVID + Sin_year + 
                        Temp_mean + Sunspots + K_Sum + Dst_var + Age*Sex,
                               data = brain_data_long_transform, family = poisson)
poisson_m10 <- glm(EMS ~ Date + Day_Week + DayOff + COVID + Sin_year + 
                        Temp_mean + Sunspots + K_Sum + Dst_level + Age*Sex,
                               data = brain_data_long_transform, family = poisson)
anova(poisson_m7, poisson_m8, test = "LR")
anova(poisson_m7, poisson_m9, test = "LR")
anova(poisson_m7, poisson_m10, test = "LR")

# Sudden_Start (-)
poisson_m11 <- glm(EMS ~ Date + Day_Week + DayOff + COVID + Sin_year + 
                        Temp_mean + Sunspots + K_Sum + Dst_var + Sudden_Start + Age*Sex,
                               data = brain_data_long_transform, family = poisson)
anova(poisson_m9, poisson_m11, test = "LR")

```

## Sunspots

Solar activity was additionally assessed by the absolute difference between Sunspots and the average Sunspots value over the 24th solar cycle. No significant difference was found between this model and the model with Sunspots.

```{r Sunspots}

brain_data %>% 
  filter(between(Date, as.Date("2008-12-15"),  as.Date("2019-12-15"))) %>% 
  pull(Sunspots) %>% mean() -> Solar_cycle_mean

brain_data_long_transform <- brain_data_long_transform %>% 
  mutate(Sunspots_dev = abs(Sunspots - mean(Solar_cycle_mean)))

poisson_m12 <- glm(EMS ~ Date + Day_Week + DayOff + COVID + Sin_year + 
                        Temp_mean +  Sunspots_dev + K_Sum + Dst_var + Age*Sex,
                               data = brain_data_long_transform, family = poisson)

AIC(poisson_m9, poisson_m12)

```

## Climatic

When additional climate data were included, ANOVA indicated that the variable **Wind** is significant for the model, while precipitation did not show a significant effect.

```{r Climatic}

poisson_m13 <- glm(EMS ~ Date + Day_Week + DayOff + COVID + Sin_year + 
                        H2O + Temp_mean + Sunspots + K_Sum + Dst_var + Age*Sex,
                               data = brain_data_long_transform, family = poisson)

poisson_m14 <- glm(EMS ~ Date + Day_Week + DayOff + COVID + Sin_year + 
                        Wind + Temp_mean + Sunspots + K_Sum + Dst_var + Age*Sex,
                               data = brain_data_long_transform, family = poisson)

anova(poisson_m9, poisson_m13, test = "LR")
anova(poisson_m9, poisson_m14, test = "LR")

poisson_m14 %>% summary()

```

## Interactions

Model interaction checks revealed, in addition to **Age-Sex**, significant effects for the following pairs: **Age-DayOff, Age-COVID, Age-Temperature, Sex-COVID, Sex-Temperature**.

```{r Interactions}

# Age Interactions
poisson_m15 <- glm(EMS ~ Date + Day_Week + Sin_year +
                        Age*DayOff + Age*COVID + Age*Sex + 
                        Age*Wind + Age*Temp_mean + Age*Sunspots + Age*K_Sum + Age*Dst_var,
                   data = brain_data_long_transform, family = poisson)
anova(poisson_m14, poisson_m15 , test = "LR") 
drop1(poisson_m15, test = "LR") 

poisson_m16 <- glm(EMS ~ Date + Day_Week + Sin_year +
                        Age*DayOff + Age*COVID + Age*Sex + Age*Temp_mean +
                        Wind + Sunspots + K_Sum + Dst_var,
                   data = brain_data_long_transform, family = poisson)
anova(poisson_m14, poisson_m16 , test = "LR") 
drop1(poisson_m16, test = "LR") 

# Sex Interactions
poisson_m17 <- glm(EMS ~ Date + Day_Week + Sin_year +
                        Age*Sex + Age*DayOff + Age*COVID + Age*Temp_mean + 
                                  Sex*DayOff + Sex*COVID + Sex*Temp_mean + 
                        Sex*Wind + Sex*Sunspots + Sex*K_Sum + Sex*Dst_var,
                               data = brain_data_long_transform, family = poisson)
anova(poisson_m16, poisson_m17 , test = "LR") 
drop1(poisson_m17, test = "LR") 


poisson_final <- glm(EMS ~ Date + Day_Week + Sin_year + 
                       Age*DayOff + Age*COVID + Age*Temp_mean + 
                       Age*Sex + Sex*COVID + Sex*Temp_mean + 
                       Wind + Sunspots + K_Sum + Dst_var, 
                     data = brain_data_long_transform, family = poisson)

anova(poisson_m16, poisson_final , test = "LR") 
drop1(poisson_final, test = "LR") 

```

## Additional Interaction

When testing the interaction of age and gender with geomagnetic variables **(K and Dst indices)** and the **number of sunspots**, no significant effects were found. The final model remains unchanged.

```{r Check}

poisson_m18 <- glm(EMS ~ Date + Day_Week + Sin_year + 
                       Age*DayOff + Age*COVID + Age*Temp_mean + 
                       Age*Sex + Sex*COVID + Sex*Temp_mean + 
                       Wind + K_Sum + Dst_var + Age*Sunspots, 
                     data = brain_data_long_transform, family = poisson)
anova(poisson_final, poisson_m18 , test = "LR") 

poisson_m19 <- glm(EMS ~ Date + Day_Week + Sin_year + 
                       Age*DayOff + Age*COVID + Age*Temp_mean + 
                       Age*Sex + Sex*COVID + Sex*Temp_mean + 
                       Wind + Sunspots + Dst_var + Age*K_Sum, 
                     data = brain_data_long_transform, family = poisson)
anova(poisson_final, poisson_m19 , test = "LR") 

poisson_m20 <- glm(EMS ~ Date + Day_Week + Sin_year + 
                       Age*DayOff + Age*COVID + Age*Temp_mean + 
                       Age*Sex + Sex*COVID + Sex*Temp_mean + 
                       Wind + Sunspots + K_Sum + Age*Dst_var, 
                     data = brain_data_long_transform, family = poisson)
anova(poisson_final, poisson_m20 , test = "LR")

poisson_m21 <- glm(EMS ~ Date + Day_Week + Sin_year + 
                       Age*DayOff + Age*COVID + Age*Temp_mean + 
                       Age*Sex + Sex*COVID + Sex*Temp_mean + 
                       Wind + K_Sum + Dst_var + Sex*Sunspots, 
                     data = brain_data_long_transform, family = poisson)
anova(poisson_final, poisson_m21 , test = "LR")

poisson_m22 <- glm(EMS ~ Date + Day_Week + Sin_year + 
                       Age*DayOff + Age*COVID + Age*Temp_mean + 
                       Age*Sex + Sex*COVID + Sex*Temp_mean + 
                       Wind + Sunspots + Dst_var + Sex*K_Sum, 
                     data = brain_data_long_transform, family = poisson)
anova(poisson_final, poisson_m22 , test = "LR")

poisson_m23 <- glm(EMS ~ Date + Day_Week + Sin_year + 
                       Age*DayOff + Age*COVID + Age*Temp_mean + 
                       Age*Sex + Sex*COVID + Sex*Temp_mean + 
                       Wind + Sunspots + K_Sum + Sex*Dst_var, 
                     data = brain_data_long_transform, family = poisson)
anova(poisson_final, poisson_m23 , test = "LR")

```

## Final Model

Diagnostic plots confirm the correctness of the final model.

```{r Final1, fig.height=10, fig.width=25}

# Summary
poisson_final %>% summary()

# Plots
poisson_final %>% autoplot() + theme_custom


```

Overdispersion decreases from 1.54 to 1.14, and autocorrelation also decreases

```{r Final2, fig.height=10, fig.width=25}
# Overdispersion 
dispersiontest(poisson_m0) 
dispersiontest(poisson_m2)
dispersiontest(poisson_final) 

# Autocorrelation of residuals
ggarrange (
residuals(poisson_m0, type = "response") %>% acf() %>% autoplot() + 
  ggtitle("Basic Model") + theme_custom,  
residuals(poisson_m2, type = "response") %>% acf() %>% autoplot() + 
  ggtitle("Long Data Model") + theme_custom,  
residuals(poisson_final, type = "response") %>% acf() %>% autoplot() + 
  ggtitle("Final Model") + theme_custom, 
nrow = 1
)

```

## Impact of Model Choice

Let's return to the negative binomial regression and the use of robust standard errors for the Poisson regression on the final set of variables.

```{r alter2, eval=FALSE, fig.width=14, include=FALSE}

# Poisson regression
tidy(poisson_final, conf.level = 0.95, conf.int = TRUE) %>%
  mutate(
    model = "Poisson") -> Poisson_data_final

# Heteroscedasticity-consistent estimation  (vcovHC)
bind_cols(
  tidy(coeftest(poisson_final, vcov. = vcovHC(poisson_final, type = "HC3"))),
  coefci(poisson_final, vcov. = vcovHC(poisson_final, type = "HC3")) %>% 
    as_tibble() %>% rename(conf.low = "2.5 %", conf.high = "97.5 %")) %>%
  mutate(
    model = "Poisson_vcovHC") -> Poisson_vcovHC_data_final

# Heteroscedasticity and Autocorrelation Consistent estimation (vcovHAC)
bind_cols(
  tidy(coeftest(poisson_final, vcov. = vcovHAC(poisson_final))),
  coefci(poisson_final, vcov. = vcovHAC(poisson_final)) %>% 
    as_tibble() %>% rename(conf.low = "2.5 %", conf.high = "97.5 %")) %>%
  mutate(
    model = "Poisson_vcovHAC") -> Poisson_vcovHAC_data_final

# Negative binomial regression
NBR_final <- glm.nb(EMS ~ Date + Day_Week + Sin_year + 
                       Age*DayOff + Age*COVID + Age*Temp_mean + 
                       Age*Sex + Sex*COVID + Sex*Temp_mean + 
                       Wind + Sunspots + K_Sum + Dst_var, 
                     data = brain_data_long_transform, link = "log")
  
tidy(NBR_final, conf.level = 0.95, conf.int = TRUE) %>% 
  mutate(
    model = "Neg_binomial") -> Neg_binomial_data_final

```

The use of robust standard errors (vcovHC/vcovHAC) and negative binomial regression in the final model, compared to the baseline, has virtually no effect on the confidence intervals for the estimates, meaning that the use of Poisson regression was justified!

```{r compare2, eval=FALSE, fig.height=15, fig.width=15, include=FALSE}

bind_rows(Poisson_data_final, Poisson_vcovHC_data_final, 
          Poisson_vcovHAC_data_final, Neg_binomial_data_final) %>% 
  filter(between(estimate, -0.2, 0.2)) %>%
  mutate(
         model = fct_recode(model, `Poisson (vcovHC)` = "Poisson_vcovHC", 
                           `Poisson (vcovHAC)`= "Poisson_vcovHAC", 
                           `Negative Binomial Regression` = "Neg_binomial")) %>% 
  ggplot(aes(x = term,y = estimate, ymin = conf.low, ymax = conf.high, colour = model))+
  geom_pointrange(position = position_dodge(width = 0.6)) +
  geom_hline(yintercept = 0, linetype = "dashed", color = "red") + 
  scale_color_brewer(palette = "Dark2", direction = -1)+
  coord_flip() +
  labs( x = "", y = "Coefficient Estimates", title = "Model Comparison", color = "Model") +
  theme_custom

```

## EMS effect

```{r effect}
final_table <- tidy(poisson_final, conf.level = 0.95, conf.int = TRUE)
```

The table below presents the factors significantly associated with ambulance calls, derived from the final model.

```{r effect_table}

final_head <- final_table %>%  
  filter(p.value < 0.05, term != "(Intercept)") %>% 
  transmute(
      Term = term, estimate = (exp(estimate) - 1) * 100, 
      conf.low = (exp(conf.low) - 1) * 100, conf.high = (exp(conf.high) - 1) * 100) %>% 
  head(12) %>% 
  gt() %>%
  fmt_number(columns = 2:4, decimals = 2) %>%  
  tab_header(
    title = md("**Factors Associated with Ambulance Calls**"),
    subtitle = md("*Only significant associations (p < 0.05)*") ) %>%
  cols_label(
    Term = md("**Variable**"), estimate = md("*Estimate*"),
    conf.low = md("*Lower CI*"), conf.high = md("*Upper CI*") ) %>% 
  tab_spanner(
    label = md("**Association, %**"),
    columns = c(estimate, conf.low, conf.high) ) %>%
  tab_options( row.striping.include_table_body = TRUE )


final_tail <- final_table %>%  
  filter(p.value < 0.05, term != "(Intercept)") %>% 
  transmute(
      Term = term, estimate = (exp(estimate) - 1) * 100, 
      conf.low = (exp(conf.low) - 1) * 100, conf.high = (exp(conf.high) - 1) * 100) %>% 
  tail(12) %>% 
  gt() %>%
  fmt_number(columns = 2:4, decimals = 2) %>%  
  tab_header(
    title = md("**Factors Associated with Ambulance Calls**"),
    subtitle = md("*Only significant associations (p < 0.05)*") ) %>%
  cols_label(
    Term = md("**Variable**"), estimate = md("*Estimate*"),
    conf.low = md("*Lower CI*"), conf.high = md("*Upper CI*") ) %>% 
  tab_spanner(
    label = md("**Association, %**"),
    columns = c(estimate, conf.low, conf.high) ) %>%
  tab_options( row.striping.include_table_body = TRUE )


gt_group(final_head, final_tail)


```


## Temperature fluctuations

New variables added to the dataset:
- Temp_var – daily temperature range.
- Temp_diff – temperature difference between days.
```{r temp_fluc}

brain_data_long_transform <- brain_data_long_transform %>% 
  mutate(Temp_var = Temp_max - Temp_min, 
         Temp_diff = (Temp_mean - lag(Temp_mean, n = 6, default = first(Temp_mean))) %>% abs(), 
         .after = COVID) 

```

ANOVA results indicate that the daily temperature range has a significant impact on the model, while wind is no longer a significant factor:
```{r temp_var}

# Temperature between days (-)
poisson_m24 <-  glm(EMS ~ Date + Day_Week + Sin_year + 
                       Age*DayOff + Age*COVID + Age*Temp_mean + 
                       Age*Sex + Sex*COVID + Sex*Temp_mean + 
                       Wind + Sunspots + K_Sum + Dst_var + Temp_diff,
                     data = brain_data_long_transform, family = poisson)

anova(poisson_final, poisson_m24 , test = "LR") 
drop1(poisson_m24, test = "LR")


# Temperature during the day (+), Interactions (-)
poisson_m25 <- glm(EMS ~ Date + Day_Week + Sin_year + 
                       Age*DayOff + Age*COVID + Age*Temp_mean + 
                       Age*Sex + Sex*COVID + Sex*Temp_mean + 
                       Wind + Sunspots + K_Sum + Dst_var + Temp_var,
                     data = brain_data_long_transform, family = poisson)

anova(poisson_final, poisson_m25 , test = "LR") 
drop1(poisson_m25, test = "LR") # +Temp_var, - Wind       

poisson_m26 <- glm(EMS ~ Date + Day_Week + Sin_year + 
                       Age*DayOff + Age*COVID + Age*Temp_mean + 
                       Age*Sex + Sex*COVID + Sex*Temp_mean + 
                       Wind + Sunspots + K_Sum + Dst_var + 
                       Age*Temp_var + Sex*Temp_var,
                     data = brain_data_long_transform, family = poisson)

anova(poisson_m25, poisson_m26 , test = "LR") # (-)

```

## Ap index

The addition of the Ap index is significant according to ANOVA, but Dst_var loses its influence.
```{r Ap}

poisson_temp <- glm(EMS ~ Date + Day_Week + Sin_year + 
                       Age*DayOff + Age*COVID + Age*Temp_mean + 
                       Age*Sex + Sex*COVID + Sex*Temp_mean + 
                      Sunspots + K_Sum + Dst_var + Temp_var,
                     data = brain_data_long_transform, family = poisson)
drop1(poisson_temp, test = "LR")

# - K_Sum + Ap_mean
poisson_m27 <- glm(EMS ~ Date + Day_Week + Sin_year + 
                       Age*DayOff + Age*COVID + Age*Temp_mean + 
                       Age*Sex + Sex*COVID + Sex*Temp_mean + 
                      Sunspots + Ap_mean + Dst_var + Temp_var,
                     data = brain_data_long_transform, family = poisson)

AIC(poisson_temp, poisson_m27)


# + Ap_mean
poisson_m28 <- glm(EMS ~ Date + Day_Week + Sin_year + 
                       Age*DayOff + Age*COVID + Age*Temp_mean + 
                       Age*Sex + Sex*COVID + Sex*Temp_mean + 
                      Sunspots + K_Sum + Ap_mean + Dst_var + Temp_var,
                     data = brain_data_long_transform, family = poisson)

anova(poisson_temp, poisson_m28 , test = "LR") 
drop1(poisson_m28, test = "LR")

```

Adding Ap_mean and Temp_var while removing Wind and Dst_var slightly improved the model.
```{r 20250223}
# -(Wind, Dst_var) + (Ap_mean + Temp_var)
poisson_20250223 <- glm(EMS ~ Date + Day_Week + Sin_year + 
                       Age*DayOff + Age*COVID + Age*Temp_mean + 
                       Age*Sex + Sex*COVID + Sex*Temp_mean + 
                      Sunspots + K_Sum + Ap_mean + Temp_var,
                     data = brain_data_long_transform, family = poisson)
anova(poisson_final,  poisson_20250223)
drop1(poisson_20250223, test = "LR")
```
